+++ src/Matrix.cuh
@@ -27,7 +27,7 @@ public:
 	__device__ __host__ float	M(int, int) const;	// ¬s—ñ®
 	__device__ __host__ Matrix	Inv(void);				// ‹ts—ñ
 	//void	print() const;
-	__device__ __host__ double get_data(int gyo, int retu) const { return m_data[gyo - 1][retu - 1]; }
+	__device__ __host__ float get_data(int gyo, int retu) const { return m_data[gyo - 1][retu - 1]; }
 	__device__ __host__ void set_data(int gyo, int retu, float value);
 };
